+++ src/tests.rs
@@ -2596,7 +2596,7 @@ fn custom_video_encoder_factory_create_and_encode_calls_callbacks() {
             &mut self,
             env: EnvironmentRef<'_>,
             format: SdpVideoFormatRef<'_>,
-        ) -> Option<Box<dyn VideoEncoderHandler>> {
+        ) -> Option<VideoEncoder> {
             assert!(!env.as_ptr().is_null());
             assert_eq!(
                 format
@@ -2608,7 +2608,9 @@ fn custom_video_encoder_factory_create_and_encode_calls_callbacks() {
                 return None;
             }
             self.created = true;
-            Some(Box::new(TestVideoEncoderHandler { encode_count: 0 }))
+            Some(VideoEncoder::new_with_handler(Box::new(
+                TestVideoEncoderHandler { encode_count: 0 },
+            )))
         }
     }
 
@@ -3020,7 +3022,7 @@ fn video_encoder_factory_create_calls_create_callback() {
             &mut self,
             env: EnvironmentRef<'_>,
             format: SdpVideoFormatRef<'_>,
-        ) -> Option<Box<dyn VideoEncoderHandler>> {
+        ) -> Option<VideoEncoder> {
             self.called.store(true, std::sync::atomic::Ordering::SeqCst);
             assert!(!env.as_ptr().is_null());
             assert_eq!(
@@ -3029,7 +3031,7 @@ fn video_encoder_factory_create_calls_create_callback() {
                     .expect("SdpVideoFormatRef::name に失敗しました"),
                 "H264"
             );
-            Some(Box::new(NoopHandler))
+            Some(VideoEncoder::new_with_handler(Box::new(NoopHandler)))
         }
     }
 
@@ -3057,7 +3059,7 @@ fn video_decoder_factory_create_calls_create_callback() {
             &mut self,
             env: EnvironmentRef<'_>,
             format: SdpVideoFormatRef<'_>,
-        ) -> Option<Box<dyn VideoDecoderHandler>> {
+        ) -> Option<VideoDecoder> {
             self.called.store(true, std::sync::atomic::Ordering::SeqCst);
             assert!(!env.as_ptr().is_null());
             assert_eq!(
@@ -3066,7 +3068,7 @@ fn video_decoder_factory_create_calls_create_callback() {
                     .expect("SdpVideoFormatRef::name に失敗しました"),
                 "H264"
             );
-            Some(Box::new(NoopHandler))
+            Some(VideoDecoder::new_with_handler(Box::new(NoopHandler)))
         }
     }
 
@@ -3274,13 +3276,15 @@ fn custom_video_decoder_factory_create_and_decode_calls_callbacks() {
             &mut self,
             env: EnvironmentRef<'_>,
             _format: SdpVideoFormatRef<'_>,
-        ) -> Option<Box<dyn VideoDecoderHandler>> {
+        ) -> Option<VideoDecoder> {
             assert!(!env.as_ptr().is_null());
             if self.created {
                 return None;
             }
             self.created = true;
-            Some(Box::new(TestVideoDecoderHandler { decode_count: 0 }))
+            Some(VideoDecoder::new_with_handler(Box::new(
+                TestVideoDecoderHandler { decode_count: 0 },
+            )))
         }
     }