+++ CHANGES.md
@@ -12,6 +12,10 @@
 
 ## develop
 
+- [CHANGE] `VideoEncoderFactoryHandler` / `VideoDecoderFactoryHandler` の create 戻り値の型を変更する
+  - `VideoEncoderFactoryHandler::create` の戻り値を `Option<VideoEncoder>` に変更する
+  - `VideoDecoderFactoryHandler::create` の戻り値を `Option<VideoDecoder>` に変更する
+  - @melpon
 - [ADD] `I420Buffer` / `NV12Buffer` に stride 指定生成と連続領域参照 API を追加する
   - C API `webrtc_I420Buffer_CreateWithStrides` / `webrtc_NV12Buffer_CreateWithStrides` を追加し、Rust API `new_with_strides` を利用できるようにする
   - Rust API `data` / `data_mut` を追加し、YUV プレーン全体を連続メモリとして扱えるようにする